+++ CHANGES.md
@@ -4,6 +4,7 @@
 Changes in 3.4.1
 ----------------
 - Add maxlength & length to Fees & Payments text inputs in Student_Billing/functions.inc.php
+- Add maxlength & length to Salaries & Payments text inputs in Accounting/functions.inc.php
 
 Changes in 3.4
 --------------